modification controllers

+++ controllers/sauces.js
@@ -16,6 +16,7 @@ exports.createSauce = (req, res, next) => {
         .catch(error => res.status(400).json({ error }));
 };
 
+
 // Mofifier une sauce
 exports.modifySauce = (req, res, next) => {
     const sauceObject = req.file ? {
@@ -24,6 +25,7 @@ exports.modifySauce = (req, res, next) => {
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     } : { ...req.body }
 
+    delete sauceObject.userId
     Sauce.findOne({ _id: req.params.id })
         .then(sauce => {
             if (req.auth.userId !== sauce.userId) {
@@ -31,16 +33,28 @@ exports.modifySauce = (req, res, next) => {
 
             } else {
                 const filename = sauce.imageUrl.split('/images/')[1]
-                fs.unlink(`images/${filename}`, () => {
                 Sauce.updateOne({ _id: req.params.id },
                     { ...sauceObject, _id: req.params.id },
                 )
                     .then(() => res.status(201).json({ message: "modified sauce" }))
+
+                    // Condition afin de vérifier si l'image a été modifiée
+                    .then(sauce => {
+                        if (sauce.imageUrl !== req.body.sauce) {
+                            fs.unlink(`images/${filename}`, () => {
+                                console.log("old picture deleted")
+                            })
+                        }else {
+                            console.log("preserved image")
+                        }
+
+                    })
                     .catch(error => res.status(400).json({ error }));
-                })
             }
-        })
-}
+        }) 
+        .catch(error => res.status(500).json({ error }))
+};
+
 
 // Supprimer une sauce
 exports.deleteSauce = (req, res, next) => {
@@ -51,9 +65,9 @@ exports.deleteSauce = (req, res, next) => {
 
                 const filename = sauce.imageUrl.split('/images/')[1]
                 fs.unlink(`images/${filename}`, () => {
-                Sauce.deleteOne({ _id: req.params.id })
-                    .then(() => res.status(200).json({ message: "Sauce removed" }))      
-                    .catch(error => res.status(400).json({ error }));
+                    Sauce.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: "Sauce removed" }))
+                        .catch(error => res.status(400).json({ error }));
                 })
             } else {
                 return res.status(403).json({ message: "unauthorized request" })
@@ -86,6 +100,7 @@ exports.likesAndDislikeSauce = (req, res, next) => {
         .then(sauce => {
 
             switch (like) {
+                // L'utilisateur aime une sauce
                 case 1:
                     if (!sauce.usersLiked.includes(userId)) {
                         Sauce.updateOne({ _id: req.params.id },
@@ -98,6 +113,7 @@ exports.likesAndDislikeSauce = (req, res, next) => {
                     }
                     break;
 
+                // L'utilisateur n'aime pas une sauce
                 case -1:
                     if (!sauce.usersDisliked.includes(userId)) {
                         Sauce.updateOne({ _id: req.params.id },
@@ -111,6 +127,7 @@ exports.likesAndDislikeSauce = (req, res, next) => {
                     }
                     break;
 
+                // L'utilisateur enlève son like ou son dislike
                 case 0:
                     if (sauce.usersLiked.includes(userId)) {
                         Sauce.updateOne({ _id: req.params.id },
